chore: rename fromWeb3DataEvent to fromWeb3ContractEvent
